Moved OpState delegate in the server header file.
