+++ examples/dishwasher-app/dishwasher-common/include/operational-state-delegate-impl.h
@@ -19,7 +19,7 @@
 #pragma once
 
 #include <app-common/zap-generated/cluster-objects.h>
-#include <app/clusters/operational-state-server/operational-state-delegate.h>
+#include <app/clusters/operational-state-server/operational-state-server.h>
 #include <app/util/af-enums.h>
 #include <protocols/interaction_model/StatusCode.h>
